DPL: refactor plugin manager
- Move PluginManager to a separate header so that we do not include the
  code to load plugin in the plugin declarations.
- Make the PluginManager code generic to use it for the upcoming
  metadata discovery.
