+++ Framework/Core/CMakeLists.txt
@@ -109,7 +109,7 @@ o2_add_library(Framework
                        src/OptionsHelpers.cxx
                        src/PropertyTreeHelpers.cxx
                        src/ProcessingContext.cxx
-                       src/Plugins.cxx
+                       src/PluginManager.cxx
                        src/RateLimiter.cxx
                        src/ResourcesMonitoringHelper.cxx
                        src/ResourcePolicy.cxx
